Extract utilities and constants; Relocate types
