+++ src/utilities/constants.ts
@@ -1,0 +1,24 @@
+export const AWS_ACCESS_KEY = 'AKIA4INYV4MFL7TXGS6F';
+export const AWS_SECRET = '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';
+export const expressionChannelOptions = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16];
+export const footswitchFunctions = ['Activate Preset', 'Bank Up', 'Bank Down', 'Preset Up', 'Preset Down', 'Tap: Midi Clock', 'Tap: Utility + Midi Clock'];
+export const markdownInstructions = 
+`Firmware Update
+======================================
+
+
+
+ **Directions**
+  * Select "Update From File" to select a specific local update file **OR**
+  * Select "Update From The Web" to update to the latest version
+  * The release notes will appear here
+  * To update, press the "Update" button
+  * To quit, press the "Cancel" button
+
+*If no release notes are displayed, no release notes were found for this release.*`
+export const messageDelay = 100;
+export const mfxId1 = 0x00;
+export const mfxId2 = 0x02;
+export const mfxId3 = 0x30;
+export const midiInputChannelOptions = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', 'Off'];
+export const utilityModeOptions = ['Tap', 'Switch: tip', 'Switch: ring', 'Switch: both'];
